requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
django-rest-framework==0.1.0
django-simple-history==3.8.0
djangorestframework==3.16.0
gevent==25.5.1
greenlet==3.2.3
kombu==5.5.4
mysqlclient==2.2.7
packaging==25.0
phonenumberslite==9.0.5
pillow==11.2.1
prompt_toolkit==3.0.51
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.1.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.2
